Refine Rat Logic
Seems there's still some logic happening in A_HDChase that prevents mobs from "attacking" corpses, which makes sense for most mobs, but for stuff like this it's preventing the rat from being the scavenging little pest that it is.  Will have to investigate overriding the base behavior somehow, even if it means completely re-implementing the default chase behavior for the sake of immersion.

For now, they seem to properly target corpses (thanks to ValidTarget now being virtual) and flee from alive players/mobs, so they can be left as-is for now.